Ahmed branch

+++ DEPI-PROJECT.BLL/Services/Implements/AgentService.cs
@@ -103,12 +103,12 @@ namespace DEPI_PROJECT.BLL.Services.Implements
             };
         }
 
-        public async Task<ResponseDto<bool>> UpdateAsync(AgentUpdateDto agentUpdateDto)
+        public async Task<ResponseDto<bool>> UpdateAsync(Guid AgentId, AgentUpdateDto agentUpdateDto)
         {
-            var agent = await _agentRepo.GetByIdAsync(agentUpdateDto.AgentId);
+            var agent = await _agentRepo.GetByIdAsync(AgentId);
             if (agent == null)
             {
-                throw new NotFoundException($"No agent found with ID {agentUpdateDto.AgentId}");
+                throw new NotFoundException($"No agent found with ID {AgentId}");
             }
 
             _mapper.Map<AgentUpdateDto, Agent>(agentUpdateDto, agent);
@@ -124,10 +124,16 @@ namespace DEPI_PROJECT.BLL.Services.Implements
                 IsSuccess = true
             };
         }
-        public async Task<ResponseDto<bool>> DeleteAsync(Guid AgentId)
+        public async Task<ResponseDto<bool>> DeleteAsync(Guid userId, Guid AgentId)
         {
-            bool result = await _agentRepo.DeleteAsync(AgentId);
+            var response = await _userRoleService.RemoveUserFromRoleByRoleName(new UserRoleByRoleNameDto { UserId = userId, RoleName = UserRoleOptions.Agent.ToString() });
 
+            if (!response.IsSuccess)
+            {
+                return response;
+            }
+
+            bool result = await _agentRepo.DeleteAsync(AgentId);
             if (!result)
             {
                 throw new Exception("An error occurred while deleting agent");
